requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==2.2.6
djangorestframework==3.11.0
pytz==2020.1
sqlparse==0.3.1
whitenoise==3.3.1
